Sales-Dash/requirements.txt
streamlit==1.29.0
pandas>=1.3
numpy
plotly>=5.5
python-dateutil
openpyxl
pyarrow